pyproject.toml
[project]
name = "web-midi-streamer"
version = "0.1.0"
description = "WebRTC-based MIDI streaming application"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
]
